[scide] [sclang] Make QtWebEngine optional (#4328)
* sclang: Refactor QcCallback declaration out of QcWebView

This commit relocates QcCallback out of the QcWebView header. QcCallback is not specific to WebView.

Also removes a dangerous "using" in header file.

* sclang: Refactor QtCollider CMakeLists

* sclang: Fix build with QT_WEBENGINE off

* sclang: Connect CMake to QT_WEBENGINE macro

* help: Expand docs for WebView/HelpBrowser

* sclang: Rename QT_WEBENGINE to SC_USE_WEBENGINE

* sclang: Make sure WebView headers appear in QT_COLLIDER_SRCS unconditionally

* cmake: include all C++ files in target even if no objects

* scide: Don't compile help browser code if SC_USE_WEBENGINE is off

* scide: Fix build with SC_USE_WEBENGINE

* scide: Remove some Qt components if SC_USE_WEBENGINE is off

* Change SC_USE_WEBENGINE to SC_USE_QTWEBENGINE

* scide: Remove IDE help menu item when SC_USE_QTWEBENGINE is off

* scide: Fix crashes for certain menu items and actions with QtWebEngine off

* help: Explain difference between the two help browsers

* Fix whitespace in lang/CMakeLists.txt
