+++ HelpSource/Classes/WebView.schelp
@@ -6,6 +6,8 @@ DESCRIPTION::
 
 WebView displays web pages and provides all the standard browsing functionality.
 
+Since the Qt WebEngine dependency is hefty and difficult to install on some systems, it is possible for sclang to have been built without WebView support (using the CMake flag code:: -DSC_USE_QTWEBENGINE=OFF :: at compile). If so, attempting to invoke this class will throw an error.
+
 CLASSMETHODS::
 
 PRIVATE:: key
