+++ QtCollider/style/ProxyStyle.cpp
@@ -3,9 +3,12 @@
 #include "../QcApplication.h"
 
 #include <QtGlobal>
-#include <QWebEngineView>
 #include <QStyleOptionSlider>
 #include <QPainter>
+
+#ifdef SC_USE_QTWEBENGINE
+#    include <QWebEngineView>
+#endif
 
 #ifdef Q_OS_MAC
 #    include "../hacks/hacks_mac.hpp"
@@ -30,6 +33,7 @@ static bool AlwaysShowScrollbars() {
 
 void ProxyStyle::drawComplexControl(ComplexControl ctrl, const QStyleOptionComplex* opt, QPainter* p,
                                     const QWidget* w) const {
+#ifdef SC_USE_QTWEBENGINE
     // FIXME: this is a workaround for the WebKit bug #104116 (or a variation on it).
     if (ctrl == QStyle::CC_ScrollBar && qobject_cast<const QWebEngineView*>(w) != 0
         && opt->type == QStyleOption::SO_Slider) {
@@ -41,6 +45,7 @@ void ProxyStyle::drawComplexControl(ComplexControl ctrl, const QStyleOptionCompl
         QProxyStyle::drawComplexControl(ctrl, &opt2, p, w);
         return;
     }
+#endif // SC_USE_QTWEBENGINE
 
     if (ctrl == QStyle::CC_ScrollBar && AlwaysShowScrollbars()) {
         const QStyleOptionSlider* optSlider = static_cast<const QStyleOptionSlider*>(opt);
