+++ editors/sc-ide/core/main.cpp
@@ -23,13 +23,9 @@
 #include "session_manager.hpp"
 #include "util/standard_dirs.hpp"
 #include "../widgets/main_window.hpp"
-#include "../widgets/help_browser.hpp"
 #include "../widgets/lookup_dialog.hpp"
 #include "../widgets/code_editor/highlighter.hpp"
 #include "../widgets/style/style.hpp"
-#include "../widgets/util/WebSocketClientWrapper.hpp"
-#include "../widgets/util/WebSocketTransport.hpp"
-#include "../widgets/util/IDEWebChannelWrapper.hpp"
 #include "../../../QtCollider/hacks/hacks_mac.hpp"
 #include "../primitives/localsocket_utils.hpp"
 
@@ -45,8 +41,14 @@
 #include <QLibraryInfo>
 #include <QTranslator>
 #include <QDebug>
-#include <QWebChannel>
 #include <QStyleFactory>
+
+#ifdef SC_USE_QTWEBENGINE
+#    include <QWebChannel>
+#    include "../widgets/util/WebSocketClientWrapper.hpp"
+#    include "../widgets/util/WebSocketTransport.hpp"
+#    include "../widgets/util/IDEWebChannelWrapper.hpp"
+#endif // SC_USE_QTWEBENGINE
 
 using namespace ScIDE;
 
@@ -129,6 +131,7 @@ int main(int argc, char* argv[]) {
     if (startInterpreter)
         main->scProcess()->startLanguage();
 
+#ifdef SC_USE_QTWEBENGINE
     // setup HelpBrowser server
     QWebSocketServer server("SCIDE HelpBrowser Server", QWebSocketServer::NonSecureMode);
     if (!server.listen(QHostAddress::LocalHost, 12344)) {
@@ -144,6 +147,7 @@ int main(int argc, char* argv[]) {
     // publish IDE interface
     IDEWebChannelWrapper ideWrapper { win->helpBrowserDocklet()->browser() };
     channel.registerObject("IDE", &ideWrapper);
+#endif // SC_USE_QTWEBENGINE
 
     return app.exec();
 }
@@ -292,6 +296,7 @@ bool Main::nativeEventFilter(const QByteArray&, void* message, long*) {
 }
 
 bool Main::openDocumentation(const QString& string) {
+#ifdef SC_USE_QTWEBENGINE
     QString symbol = string.trimmed();
     if (symbol.isEmpty())
         return false;
@@ -300,13 +305,20 @@ bool Main::openDocumentation(const QString& string) {
     helpDock->browser()->gotoHelpFor(symbol);
     helpDock->focus();
     return true;
+#else // SC_USE_QTWEBENGINE
+    return false;
+#endif // SC_USE_QTWEBENGINE
 }
 
 bool Main::openDocumentationForMethod(const QString& className, const QString& methodName) {
+#ifdef SC_USE_QTWEBENGINE
     HelpBrowserDocklet* helpDock = MainWindow::instance()->helpBrowserDocklet();
     helpDock->browser()->gotoHelpForMethod(className, methodName);
     helpDock->focus();
     return true;
+#else // SC_USE_QTWEBENGINE
+    return false;
+#endif // SC_USE_QTWEBENGINE
 }
 
 void Main::openDefinition(const QString& string, QWidget* parent) {
