+++ editors/sc-ide/widgets/code_editor/autocompleter.cpp
@@ -28,7 +28,10 @@
 #include "../../core/main.hpp"
 #include "../../core/util/standard_dirs.hpp"
 #include "../main_window.hpp"
-#include "../help_browser.hpp"
+
+#ifdef SC_USE_QTWEBENGINE
+#    include "../help_browser.hpp"
+#endif // SC_USE_QTWEBENGINE
 
 #include <yaml-cpp/node/node.h>
 #include <yaml-cpp/parser.h>
@@ -1213,9 +1216,11 @@ void AutoCompleter::parseClassNode(DocNode* node, QString* str) {
 }
 
 void AutoCompleter::gotoHelp(QString symbol) {
+#ifdef SC_USE_QTWEBENGINE
     HelpBrowserDocklet* helpDock = MainWindow::instance()->helpBrowserDocklet();
     helpDock->browser()->gotoHelpFor(symbol);
     helpDock->focus();
+#endif // SC_USE_QTWEBENGINE
 }
 
 } // namespace ScIDE
