+++ editors/sc-ide/widgets/main_window.hpp
@@ -101,8 +101,15 @@ public:
         Help,
         HelpAboutIDE,
         ReportABug,
+
+#ifdef SC_USE_QTWEBENGINE
+        // These QtWebEngine-only actions are branched at the preprocessor
+        // level so that accidental invocations of these actions in other code
+        // are caught at compile time.
         LookupDocumentationForCursor,
         LookupDocumentation,
+#endif // SC_USE_QTWEBENGINE
+
         ShowAbout,
         ShowAboutQT,
 
@@ -120,7 +127,9 @@ public:
     void focusCodeEditor();
     bool promptSaveDocs();
 
+#ifdef SC_USE_QTWEBENGINE
     HelpBrowserDocklet* helpBrowserDocklet() { return mHelpBrowserDocklet; }
+#endif
     PostDocklet* postDocklet() { return mPostDocklet; }
 
     static MainWindow* instance() { return mInstance; }
@@ -237,7 +246,9 @@ private:
     // Docks
     PostDocklet* mPostDocklet;
     DocumentsDocklet* mDocumentsDocklet;
+#ifdef SC_USE_QTWEBENGINE
     HelpBrowserDocklet* mHelpBrowserDocklet;
+#endif
 
     QSignalMapper mCodeEvalMapper;
     DocumentsDialog* mDocDialog;
