+++ editors/sc-ide/widgets/style/style.cpp
@@ -32,7 +32,10 @@
 #include <QToolButton>
 #include <QLayout>
 #include <QDebug>
-#include <QWebEngineView>
+
+#ifdef SC_USE_QTWEBENGINE
+#    include <QWebEngineView>
+#endif // SC_USE_QTWEBENGINE
 
 namespace ScIDE {
 
@@ -69,6 +72,7 @@ void Style::drawComplexControl(ComplexControl control, const QStyleOptionComplex
     }
 
     switch (control) {
+#ifdef SC_USE_QTWEBENGINE
     // FIXME: this is a workaround for the WebKit bug #104116 (or a variation on it).
     case QStyle::CC_ScrollBar: {
         if (qobject_cast<const QWebEngineView*>(widget) != 0 && option->type == QStyleOption::SO_Slider) {
@@ -81,6 +85,7 @@ void Style::drawComplexControl(ComplexControl control, const QStyleOptionComplex
             return;
         }
     }
+#endif // SC_USE_QTWEBENGINE
     case QStyle::CC_ToolButton: {
         // TODO: We only draw either text, or icon, or arrow
